lint, CI, Readme

+++ contracts/TornadoPool.sol
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: MIT
 // https://tornado.cash
 /*
-* d888888P                                           dP              a88888b.                   dP
-*    88                                              88             d8'   `88                   88
-*    88    .d8888b. 88d888b. 88d888b. .d8888b. .d888b88 .d8888b.    88        .d8888b. .d8888b. 88d888b.
-*    88    88'  `88 88'  `88 88'  `88 88'  `88 88'  `88 88'  `88    88        88'  `88 Y8ooooo. 88'  `88
-*    88    88.  .88 88       88    88 88.  .88 88.  .88 88.  .88 dP Y8.   .88 88.  .88       88 88    88
-*    dP    `88888P' dP       dP    dP `88888P8 `88888P8 `88888P' 88  Y88888P' `88888P8 `88888P' dP    dP
-* ooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
-*/
+ * d888888P                                           dP              a88888b.                   dP
+ *    88                                              88             d8'   `88                   88
+ *    88    .d8888b. 88d888b. 88d888b. .d8888b. .d888b88 .d8888b.    88        .d8888b. .d8888b. 88d888b.
+ *    88    88'  `88 88'  `88 88'  `88 88'  `88 88'  `88 88'  `88    88        88'  `88 Y8ooooo. 88'  `88
+ *    88    88.  .88 88       88    88 88.  .88 88.  .88 88.  .88 dP Y8.   .88 88.  .88       88 88    88
+ *    dP    `88888P' dP       dP    dP `88888P8 `88888P8 `88888P' 88  Y88888P' `88888P8 `88888P' dP    dP
+ * ooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooooo
+ */
 
 pragma solidity ^0.6.0;
 pragma experimental ABIEncoderV2;
@@ -16,8 +16,9 @@ pragma experimental ABIEncoderV2;
 import "@openzeppelin/contracts/utils/ReentrancyGuard.sol"; // todo: maybe remove?
 
 interface IVerifier {
-  function verifyProof(bytes memory _proof, uint256[9] memory _input) external returns(bool);
-  function verifyProof(bytes memory _proof, uint256[23] memory _input) external returns(bool);
+  function verifyProof(bytes memory _proof, uint256[9] memory _input) external returns (bool);
+
+  function verifyProof(bytes memory _proof, uint256[23] memory _input) external returns (bool);
 }
 
 contract TornadoPool is ReentrancyGuard {
@@ -26,7 +27,7 @@ contract TornadoPool is ReentrancyGuard {
 
   mapping(bytes32 => bool) public nullifierHashes;
   bytes32 public currentRoot;
-  uint public currentCommitmentIndex;
+  uint256 public currentCommitmentIndex;
   IVerifier public verifier2;
   IVerifier public verifier16;
 
@@ -38,7 +39,7 @@ contract TornadoPool is ReentrancyGuard {
   }
 
   // todo: event Transaction();
-  event NewCommitment(bytes32 commitment, uint index, bytes encryptedOutput);
+  event NewCommitment(bytes32 commitment, uint256 index, bytes encryptedOutput);
   event NewNullifier(bytes32 nullifier);
   event Withdraw(bytes32 indexed nullifier); // todo emit it on withdraw so we can easily find the withdraw tx for user on UI
 
@@ -47,7 +48,11 @@ contract TornadoPool is ReentrancyGuard {
     @param _verifier2 the address of SNARK verifier for this contract
     @param _verifier16 the address of SNARK verifier for this contract
   */
-  constructor(IVerifier _verifier2, IVerifier _verifier16, bytes32 _currentRoot) public {
+  constructor(
+    IVerifier _verifier2,
+    IVerifier _verifier16,
+    bytes32 _currentRoot
+  ) public {
     verifier2 = _verifier2;
     verifier16 = _verifier16;
     currentRoot = _currentRoot;
@@ -63,58 +68,68 @@ contract TornadoPool is ReentrancyGuard {
     uint256 _fee,
     ExtData calldata _extData,
     bytes32 _extDataHash
-  )
-    external payable nonReentrant
-  {
+  ) external payable nonReentrant {
     require(currentRoot == _root, "Invalid merkle root");
-    for(uint256 i = 0; i < _inputNullifiers.length; i++) {
+    for (uint256 i = 0; i < _inputNullifiers.length; i++) {
       require(!isSpent(_inputNullifiers[i]), "Input is already spent");
     }
     require(uint256(_extDataHash) == uint256(keccak256(abi.encode(_extData))) % FIELD_SIZE, "Incorrect external data hash");
     if (_inputNullifiers.length == 2) {
-      require(verifier2.verifyProof(_proof, [
-        uint256(_root),
-        uint256(_newRoot),
-        uint256(_inputNullifiers[0]),
-        uint256(_inputNullifiers[1]),
-        uint256(_outputCommitments[0]),
-        uint256(_outputCommitments[1]),
-        _extAmount,
-        _fee,
-        uint256(_extDataHash)
-      ]), "Invalid transaction proof");
+      require(
+        verifier2.verifyProof(
+          _proof,
+          [
+            uint256(_root),
+            uint256(_newRoot),
+            uint256(_inputNullifiers[0]),
+            uint256(_inputNullifiers[1]),
+            uint256(_outputCommitments[0]),
+            uint256(_outputCommitments[1]),
+            _extAmount,
+            _fee,
+            uint256(_extDataHash)
+          ]
+        ),
+        "Invalid transaction proof"
+      );
     } else if (_inputNullifiers.length == 16) {
-      require(verifier16.verifyProof(_proof, [
-        uint256(_root),
-        uint256(_newRoot),
-        uint256(_inputNullifiers[0]),
-        uint256(_inputNullifiers[1]),
-        uint256(_inputNullifiers[2]),
-        uint256(_inputNullifiers[3]),
-        uint256(_inputNullifiers[4]),
-        uint256(_inputNullifiers[5]),
-        uint256(_inputNullifiers[6]),
-        uint256(_inputNullifiers[7]),
-        uint256(_inputNullifiers[8]),
-        uint256(_inputNullifiers[9]),
-        uint256(_inputNullifiers[10]),
-        uint256(_inputNullifiers[11]),
-        uint256(_inputNullifiers[12]),
-        uint256(_inputNullifiers[13]),
-        uint256(_inputNullifiers[14]),
-        uint256(_inputNullifiers[15]),
-        uint256(_outputCommitments[0]),
-        uint256(_outputCommitments[1]),
-        _extAmount,
-        _fee,
-        uint256(_extDataHash)
-        ]), "Invalid transaction proof");
+      require(
+        verifier16.verifyProof(
+          _proof,
+          [
+            uint256(_root),
+            uint256(_newRoot),
+            uint256(_inputNullifiers[0]),
+            uint256(_inputNullifiers[1]),
+            uint256(_inputNullifiers[2]),
+            uint256(_inputNullifiers[3]),
+            uint256(_inputNullifiers[4]),
+            uint256(_inputNullifiers[5]),
+            uint256(_inputNullifiers[6]),
+            uint256(_inputNullifiers[7]),
+            uint256(_inputNullifiers[8]),
+            uint256(_inputNullifiers[9]),
+            uint256(_inputNullifiers[10]),
+            uint256(_inputNullifiers[11]),
+            uint256(_inputNullifiers[12]),
+            uint256(_inputNullifiers[13]),
+            uint256(_inputNullifiers[14]),
+            uint256(_inputNullifiers[15]),
+            uint256(_outputCommitments[0]),
+            uint256(_outputCommitments[1]),
+            _extAmount,
+            _fee,
+            uint256(_extDataHash)
+          ]
+        ),
+        "Invalid transaction proof"
+      );
     } else {
       revert("unsupported input count");
     }
 
     currentRoot = _newRoot;
-    for(uint256 i = 0; i < _inputNullifiers.length; i++) {
+    for (uint256 i = 0; i < _inputNullifiers.length; i++) {
       nullifierHashes[_inputNullifiers[i]] = true;
     }
 
@@ -136,12 +151,12 @@ contract TornadoPool is ReentrancyGuard {
     // todo enforce currentCommitmentIndex value in snark
     emit NewCommitment(_outputCommitments[0], currentCommitmentIndex++, _extData.encryptedOutput1);
     emit NewCommitment(_outputCommitments[1], currentCommitmentIndex++, _extData.encryptedOutput2);
-    for(uint256 i = 0; i < _inputNullifiers.length; i++) {
+    for (uint256 i = 0; i < _inputNullifiers.length; i++) {
       emit NewNullifier(_inputNullifiers[i]);
     }
   }
 
-  function calculateExternalAmount(uint256 _extAmount) public pure returns(int256) {
+  function calculateExternalAmount(uint256 _extAmount) public pure returns (int256) {
     // -MAX_EXT_AMOUNT < extAmount < MAX_EXT_AMOUNT
     if (_extAmount < MAX_EXT_AMOUNT) {
       return int256(_extAmount);
@@ -154,7 +169,7 @@ contract TornadoPool is ReentrancyGuard {
   }
 
   /** @dev whether a note is already spent */
-  function isSpent(bytes32 _nullifierHash) public view returns(bool) {
+  function isSpent(bytes32 _nullifierHash) public view returns (bool) {
     return nullifierHashes[_nullifierHash];
   }
 }
